Refactor cryostat geometry calculations into a new Cryostat class and update related modules
